requirements.txt
APScheduler==3.10.1
Django==4.2.1
djangorestframework==3.14.0
finnhub-python==2.4.18
idna==3.4
psycopg2-binary==2.9.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
